rearrange helper functions
