+++ .pre-commit-config.yaml
@@ -4,7 +4,8 @@ repos:
         hooks:
         - id: isort
           additional_dependencies: [toml]
-          exclude: examples/*
+          args: ["--profile", "black"]
+          exclude: examples/.*
       - repo: https://github.com/python/black
         rev: 22.3.0
         hooks:
