+++ tests/unit/systems/test_systems_ensemble.py
@@ -1,38 +1,38 @@
-import subprocess
 import os
-import tempfile
 import shutil
+import tempfile
 import unittest
-from feast import FeatureStore
-import numpy as np
 
-from datetime import datetime
-from merlin.models.loader.tf_utils import configure_tensorflow
-from merlin.models.utils.dataset import unique_rows_by_features
-from merlin.schema import Schema, ColumnSchema
-from merlin.systems.dag.ops.faiss import QueryFaiss, setup_faiss
+import numpy as np
+from feast import FeatureStore
+from merlin.datasets.ecommerce import transform_aliccp
+from merlin.datasets.synthetic import generate_data
+from merlin.io.dataset import Dataset
+from merlin.schema import ColumnSchema, Schema
+from merlin.schema.tags import Tags
+from merlin.systems.dag.ensemble import Ensemble
+from merlin.systems.dag.ops.faiss import QueryFaiss
 from merlin.systems.dag.ops.feast import QueryFeast
 from merlin.systems.dag.ops.tensorflow import PredictTensorflow
 from merlin.systems.dag.ops.unroll_features import UnrollFeatures
 from merlin.systems.dag.ops.workflow import TransformWorkflow
-from merlin.systems.dag.ensemble import Ensemble
 from nvtabular.ops import (
-    Categorify,
-    TagAsUserID,
-    TagAsItemID,
-    TagAsItemFeatures,
-    TagAsUserFeatures,
     AddMetadata,
+    Categorify,
     Filter,
+    TagAsItemFeatures,
+    TagAsItemID,
+    TagAsUserFeatures,
+    TagAsUserID,
 )
 from nvtabular.workflow import Workflow
-from merlin.schema.tags import Tags
 
-import merlin.models.tf as mm
-from merlin.io.dataset import Dataset
-from merlin.datasets.ecommerce import transform_aliccp
-import tensorflow as tf
-from merlin.datasets.synthetic import generate_data
+from .configuration_helpers import (
+    _configure_faiss,
+    _configure_feast,
+    _train_dlrm_model,
+    _train_two_tower,
+)
 
 os.environ["TF_GPU_ALLOCATOR"] = "cuda_malloc_async"
 
@@ -51,9 +51,6 @@ USER_FEATURES = [
     "user_categories",
 ]
 ITEM_FEATURES = ["item_category", "item_shop", "item_brand"]
-
-
-configure_tensorflow()
 
 
 class MerlinTestCase(unittest.TestCase):
@@ -302,107 +299,6 @@ class MerlinTestCase(unittest.TestCase):
         Ensemble(pipeline, request_schema)
 
 
-def _configure_feast(feature_store_path, data_to_ingest: Dataset):
-    """
-    This configures a file-based Feast feature store in the directory `feast_feature_store`, which
-    should be a subdirectory of MerlinTestCase.base_data_dir.
-
-    The steps that this takes are:
-
-    * Make the directory if it doesn't already exist.
-    * Copy `user_features.py` and `item_features.py` into the directory.
-    * Write the feature_store.yaml file into the directory.
-    * Add timestamp columns to the user and item features.
-    * Save the user and item features into the feature store directory.
-    * Run `feast apply` and `feast materialize` to load data into the online store.
-    """
-
-    if not os.path.exists(feature_store_path):
-        os.makedirs(os.path.join(feature_store_path, "data"))
-
-    # Move the required files into the feature_store_path
-    with open(
-        os.path.join(feature_store_path, "feature_store.yaml"), "w", encoding="utf8"
-    ) as config_file:
-        config_file.write(
-            """
-project: feast_feature_store
-registry: data/registry.db
-provider: local
-online_store:
-    path: data/online_store.db
-"""
-        )
-    cwd = os.path.dirname(os.path.abspath(__file__))
-    shutil.copy(
-        os.path.join(cwd, "feast_feature_definitions", "item_features.py"),
-        feature_store_path,
-    )
-    shutil.copy(
-        os.path.join(cwd, "feast_feature_definitions", "user_features.py"),
-        feature_store_path,
-    )
-
-    # Split training data into user / item and add timestamps.
-
-    user_features = (
-        unique_rows_by_features(data_to_ingest, Tags.USER, Tags.USER_ID)
-        .compute()
-        .reset_index(drop=True)
-    )
-    user_features["datetime"] = datetime.now()
-    user_features["datetime"] = user_features["datetime"].astype("datetime64[ns]")
-    user_features["created"] = datetime.now()
-    user_features["created"] = user_features["created"].astype("datetime64[ns]")
-    user_features.to_parquet(
-        os.path.join(feature_store_path, "data", "user_features.parquet")
-    )
-
-    item_features = (
-        unique_rows_by_features(data_to_ingest, Tags.ITEM, Tags.ITEM_ID)
-        .compute()
-        .reset_index(drop=True)
-    )
-    item_features["datetime"] = datetime.now()
-    item_features["datetime"] = item_features["datetime"].astype("datetime64[ns]")
-    item_features["created"] = datetime.now()
-    item_features["created"] = item_features["created"].astype("datetime64[ns]")
-    item_features.to_parquet(
-        os.path.join(feature_store_path, "data", "item_features.parquet")
-    )
-
-    # These environment variables tell `user_features.py` and `item_features.py` where the parquet
-    # files are stored, since it changes with every test.
-    os.environ.setdefault(
-        "FEAST_USER_FEATURES_PATH",
-        os.path.join(feature_store_path, "data", "user_features.parquet"),
-    )
-    os.environ.setdefault(
-        "FEAST_ITEM_FEATURES_PATH",
-        os.path.join(feature_store_path, "data", "item_features.parquet"),
-    )
-
-    # Ingest data.
-    subprocess.run(["feast", "-c", feature_store_path, "apply"], check=True)
-    subprocess.run(
-        [
-            "feast",
-            "-c",
-            feature_store_path,
-            "materialize",
-            "1995-01-01T01:01:01",
-            "2025-01-01T01:01:01",
-        ],
-        check=True,
-    )
-
-    # This is a useful place to drop a debugger and manually see what is in feast.
-    # import pdb; pdb.set_trace()
-    # import feast
-    # fs = feast.FeatureStore(feature_store_path)
-    # fs.get_online_features(["user_features:user_shops"], [{"user_id":1}]).to_dict()
-
-
 def _retrieval_workflow():
     """
     NVT workflow used to transform features for retrieval model.
@@ -437,61 +333,3 @@ def _ranking_workflow():
 
     outputs = user_id + item_id + item_features + user_features + targets
     return outputs
-
-
-def _configure_faiss(faiss_index_path: str, model_tt, item_features: Dataset):
-    item_embs = model_tt.item_embeddings(item_features, batch_size=1024)
-    item_embeddings = np.ascontiguousarray(
-        item_embs.compute(scheduler="synchronous").to_numpy()
-    )
-    setup_faiss(item_embeddings, faiss_index_path)
-
-
-def _train_two_tower(train_data_path, valid_data_path, query_tower_output_path):
-    train_tt = Dataset(os.path.join(train_data_path, "*.parquet"))
-    valid_tt = Dataset(os.path.join(valid_data_path, "*.parquet"))
-
-    schema = train_tt.schema
-
-    schema = schema.select_by_tag([Tags.ITEM_ID, Tags.USER_ID, Tags.ITEM, Tags.USER])
-    model_tt = mm.TwoTowerModel(
-        schema,
-        query_tower=mm.MLPBlock([128, 64], no_activation_last_layer=True),
-        samplers=[mm.InBatchSampler()],
-        embedding_options=mm.EmbeddingOptions(infer_embedding_sizes=True),
-    )
-    model_tt.compile(
-        optimizer="adam",
-        run_eagerly=False,
-        loss="categorical_crossentropy",
-        metrics=[mm.RecallAt(10), mm.NDCGAt(10)],
-    )
-    model_tt.fit(train_tt, validation_data=valid_tt, batch_size=1024 * 8, epochs=1)
-    query_tower = model_tt.retrieval_block.query_block()
-    query_tower.save(query_tower_output_path)
-
-    return model_tt
-
-
-def _train_dlrm_model(
-    train_data_path: str, valid_data_path: str, model_output_path: str
-):
-    # define train and valid dataset objects
-    train = Dataset(os.path.join(train_data_path, "*.parquet"), part_size="500MB")
-    valid = Dataset(os.path.join(valid_data_path, "*.parquet"), part_size="500MB")
-
-    # define schema object
-    schema: Schema = train.schema
-
-    target_column = schema.select_by_tag(Tags.TARGET).column_names[0]
-    model = mm.DLRMModel(
-        schema,
-        embedding_dim=64,
-        bottom_block=mm.MLPBlock([128, 64]),
-        top_block=mm.MLPBlock([128, 64, 32]),
-        prediction_tasks=mm.BinaryClassificationTask(target_column),
-    )
-
-    model.compile(optimizer="adam", run_eagerly=False, metrics=[tf.keras.metrics.AUC()])
-    model.fit(train, validation_data=valid, batch_size=16 * 1024)
-    model.save(model_output_path)
